Populate Home View with wireframes for posts
Add a mock model for later use when I start populating each cell from each section with mock data, rather than just filling cells with colors. Mock includes:
	- User
	- Post
	- Comments
	- No actions
Add buffer using the footerInSection method, currently displaying a blank view
Gave each cell a fixed height for their respective rows, with the exception of the primaryContent which fills the width of the tableView (and is situated between header and actions)
Rows are currently non-selectable in HV
In the cell configuration, add some logic to handle mock data for later addition
Comments are currently set to 2 rows or less in the "default" view of the HV. If more than 2 comments, it caps at 2.